Substitute QVector and QList with their std:: counterparts
I use <algorithm> and other STL features and since QVector does not
provide a significant advantage I opted for std::vector.

The main reason is that QVector in Qt5.12 does not yet provide
a constructor from begin() and end() iterators which is quite nice if
you want to access ranges.

Signed-off-by: Thomas Ruschival <thomas@ruschival.de>